+++ web/web/settings.py
@@ -231,6 +231,9 @@ ALLOWED_HOSTS = ["*"]
 # Max size
 MAX_UPLOAD_SIZE = web_cfg.general.max_sample_size
 
+# Don't forget to give some love to @doomedraven ;)
+RATELIMIT_ERROR_MSG = "You are blocked, Too many request without api access! We are researcher friendly, but if you buy a good whiskey to @doomedraven, we will be even more friendlier ;)"
+
 # Hack to import local settings.
 try:
     LOCAL_SETTINGS